Implement UWB getRangingInfo luna API
:Release Notes:
Implemented WRO-1977
Implement UWB getRangingInfo luna API as per the
new API spec

:Detailed Notes:
Implement UWB getRangingInfo luna API as per the
new API spec

:Testing Performed:
Please refer to QWT-4537

:QA Notes:
NA

:Issues Addressed:
http://clm.lge.com/issue/browse/WRO-1977

Reviewed-on: http://gpro.lge.com/c/webosose/com.webos.service.uwb/+/331782
Tested-by: <muralidhar.n@lge.com>

+++ src/UwbRangingInfo.cpp
@@ -26,7 +26,7 @@ string UwbRangingInfo::getRemoteDevAddr() const{
     return m_remoteDevAddr;
 }
 
-int32_t UwbRangingInfo::getElapsedTime() const{
+int64_t UwbRangingInfo::getElapsedTime() const{
     return mElapsedTime;
 }
 
@@ -58,7 +58,7 @@ void UwbRangingInfo::setRemoteDevAddr(string remoteDevAddr) {
      m_remoteDevAddr = remoteDevAddr;
 }
 
-void UwbRangingInfo::setElapsedTime(int32_t elapsedTime) {
+void UwbRangingInfo::setElapsedTime(int64_t elapsedTime) {
     mElapsedTime = elapsedTime;
 }
